Fix Previous Bad Commit (1077) Mac OSX 10.5 Leopard / Xcode 3 Compatibility - Builds using jam on Leopard with Xcode 3 using the 10.5 sdk - Will NOT build on Tiger. Leopard only. - Contrib packs are not updated. Jam build only for now.
git-svn-id: svn://svn.handbrake.fr/HandBrake/trunk@1078 b64f7644-9d1e-0410-96f1-a4d463321fa5

+++ macosx/PictureGLView.h
@@ -27,7 +27,10 @@
     float           fCoordY;
 
     uint8_t       * fBuffers[2];
-    unsigned long   fTextures[2];
+    /* Tiger */
+    //unsigned long   fTextures[2];
+    /* Leopard */
+    unsigned int   fTextures[2];
 
     int             fLastEffect;
     int             fAnimDuration;
