+++ macosx/PictureGLView.mm
@@ -414,9 +414,11 @@ static int GetAlignedSize( int size )
     /* Swap buffers only during the vertical retrace of the monitor.
        http://developer.apple.com/documentation/GraphicsImaging/
        Conceptual/OpenGL/chap5/chapter_5_section_44.html */
-    long params[] = { 1 };
-    CGLSetParameter( CGLGetCurrentContext(), kCGLCPSwapInterval,
-                     params );
+    /* Tiger */
+    //long params[] = { 1 };
+    /* Leopard */
+    int params[] = { 1 };
+    CGLSetParameter( CGLGetCurrentContext(), kCGLCPSwapInterval, params );
 
     if( !( anim & HB_ANIMATE_NONE ) )
     {
